other review comment

+++ source/extensions/filters/network/sni_dynamic_forward_proxy/proxy_filter.cc
@@ -34,9 +34,10 @@ Network::FilterStatus ProxyFilter::onNewConnection() {
     return Network::FilterStatus::Continue;
   }
 
-  // TODO(lizan): implement circuit breaker in SNI dynamic forward proxy
+  // TODO(lizan): implement circuit breaker in SNI dynamic forward proxy like it is in HTTP:
+  // https://github.com/envoyproxy/envoy/blob/master/source/extensions/filters/http/dynamic_forward_proxy/proxy_filter.cc#L65
 
-  uint16_t default_port = config_->port();
+  uint32_t default_port = config_->port();
 
   auto result = config_->cache().loadDnsCacheEntry(sni, default_port, *this);
